feat(folder): add folder CRUD and trash functionality
- Implement FolderController with CRUD + trash restore
- Add FolderService + IFolderService
- Add FolderIndexViewModel and FolderEditViewModel
- Add views for index, modals, and partial lists

+++ DMS.Infrastructure/IRepositories/IDocumentRepository.cs
@@ -5,8 +5,8 @@ namespace DMS.Infrastructure.IRepositories
 {
     public interface IDocumentRepository: ISortSearch<Document>, IRepository<Document>
     {
-        Task<double> GetTotalStorageAsync();
-        Task<double> GetTotalStorageByUserAsync(string userId);
+        Task<long> GetTotalStorageAsync();
+        Task<long> GetTotalStorageByUserAsync(string userId);
         Task<List<Document>> GetDocumentsByFolderIdAsync(string folderId, string userId);
         IQueryable<Document> SearchDocumentByFolderAsQueryable(string folderId, string userId, string searchName);
         IQueryable<Document> GetDocumentsByFolderIdAsQueryable(string folderId, string userId);
